Drop PuppetDB API v1 and v3 support
Puppet 6 & 7 only support v4 and older version are EOL.
