--- app/services/puppetdb_client/v1.rb
@@ -1,33 +1,0 @@
-module PuppetdbClient
-  class V1 < Base
-    # The payload is expected to be the certname of a node as a serialized JSON string.
-    def deactivate_node_payload(nodename)
-      payload = {
-        'command' => 'deactivate node',
-        'version' => 1,
-        'payload' => nodename.to_json
-      }.to_json
-      'payload=' + payload
-    end
-
-    def command_url
-      '/v3/commands'
-    end
-
-    def nodes_url
-      '/v3/nodes'
-    end
-
-    def facts_url
-      '/v3/facts'
-    end
-
-    def resources_url
-      '/v3/resources'
-    end
-
-    def query_nodes
-      super.map { |node| node['name'] }
-    end
-  end
-end
